# BUG/PERF: MultiIndex.difference losing EA dtype