--- a/asv_bench/benchmarks/multiindex_object.py
+++ b/asv_bench/benchmarks/multiindex_object.py
@@ -238,7 +238,7 @@ class SetOperations:
     params = [
         ("monotonic", "non_monotonic"),
         ("datetime", "int", "string", "ea_int"),
-        ("intersection", "union", "symmetric_difference"),
+        ("intersection", "union", "difference", "symmetric_difference"),
     ]
     param_names = ["index_structure", "dtype", "method"]