- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v1.6.0.rst` file if fixing a bug or adding a new feature.

Maintains EA dtypes and improves perf:

```
       before           after         ratio
     [71fc89cd]       [19e9ae76]
     <main>           <mi-difference>
-      29.0±0.8ms       16.7±0.9ms     0.58  multiindex_object.SetOperations.time_operation('monotonic', 'string', 'difference')
-      30.4±0.4ms       15.7±0.1ms     0.52  multiindex_object.SetOperations.time_operation('non_monotonic', 'string', 'difference')
-      26.7±0.7ms       13.0±0.5ms     0.49  multiindex_object.SetOperations.time_operation('monotonic', 'ea_int', 'difference')
-      30.1±0.3ms       13.3±0.1ms     0.44  multiindex_object.SetOperations.time_operation('non_monotonic', 'ea_int', 'difference')
-      25.1±0.5ms       8.87±0.2ms     0.35  multiindex_object.SetOperations.time_operation('monotonic', 'int', 'difference')
-      27.7±0.3ms       9.58±0.3ms     0.35  multiindex_object.SetOperations.time_operation('monotonic', 'datetime', 'difference')
-      27.5±0.4ms       9.13±0.3ms     0.33  multiindex_object.SetOperations.time_operation('non_monotonic', 'int', 'difference')
-      33.4±0.2ms       10.4±0.4ms     0.31  multiindex_object.SetOperations.time_operation('non_monotonic', 'datetime', 'difference')
```