@lukemanley 2022-09-23
Ah, sorry. I saw the other setop PRs merged and thought difference might have been missed. I’ll close this.